Fix looping background restricition notification in onboarding (EXPOSUREAPP-1962) (#950)
* added call to navigateToMain before opening the app settings

* Moved checks for background job notification to main activity

* refined background dialogs
