+++ Corona-Warn-App/src/device/java/de.rki.coronawarnapp/ui/main/MainFragment.kt
@@ -21,7 +21,6 @@ import de.rki.coronawarnapp.ui.viewmodel.SubmissionViewModel
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
 import de.rki.coronawarnapp.util.DialogHelper
 import de.rki.coronawarnapp.util.ExternalActionHelper
-import de.rki.coronawarnapp.util.PowerManagementHelper
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -72,7 +71,6 @@ class MainFragment : Fragment() {
         setContentDescription()
 
         showOneTimeTracingExplanationDialog()
-        showEnergyOptimizedExplanationDialog()
     }
 
     override fun onResume() {
@@ -216,36 +214,6 @@ class MainFragment : Fragment() {
                             },
                             {}
                         ))
-                }
-            }
-        }
-    }
-
-    private fun showEnergyOptimizedExplanationDialog() {
-
-        // check if the dialog explaining the effects of energy saver mode were already shown and if energy saver is enabled
-        if (!LocalData.energyOptimizedExplanationDialogWasShown() && !PowerManagementHelper.isIgnoringBatteryOptimizations(requireActivity())) {
-            lifecycleScope.launch {
-
-                withContext(Dispatchers.Main) {
-
-                    val dialog = DialogHelper.DialogInstance(
-                        requireActivity(),
-                        R.string.onboarding_energy_optimized_dialog_headline,
-                        R.string.onboarding_energy_optimized_dialog_body,
-                        R.string.onboarding_energy_optimized_dialog_button_positive,
-                        R.string.onboarding_energy_optimized_dialog_button_negative,
-                        false,
-                        {
-                            // go to battery optimization
-                            ExternalActionHelper.toBatteryOptimizationSettings(requireContext())
-                            LocalData.energyOptimizedExplanationDialogWasShown(true)
-                        },
-                        {
-                            // keep battery optimization enabled
-                            LocalData.energyOptimizedExplanationDialogWasShown(true)
-                        })
-                    DialogHelper.showDialog(dialog)
                 }
             }
         }
